backend/requirements.api.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pymongo>=4.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=0.19.0
tensorflow==2.18.0
numpy>=1.26.0  # Mise à jour pour être compatible avec TensorFlow 2.18.0
opencv-python-headless>=4.5.0
Pillow>=8.0.0